# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SistemaJava • Mercadinho PDV",
    "description": "Obtenha CSRF em /api/v1/auth/csrf; envie X-XSRF-TOKEN em mutações. Login retorna Bearer JWT.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
    "/api/v1/suppliers/{id}": {
      "put": {
        "tags": [
          "purchase-controller"
        ],
        "operationId": "editSupplier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/recipes/{id}": {
      "put": {
        "tags": [
          "purchase-controller"
        ],
        "operationId": "editRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "purchase-controller"
        ],
        "operationId": "deleteRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/suppliers": {
      "get": {
        "tags": [
          "purchase-controller"
        ],
        "operationId": "suppliers",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "purchase-controller"
        ],
        "operationId": "createSupplier",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/recipes": {
      "get": {
        "tags": [
          "purchase-controller"
        ],
        "operationId": "recipes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "purchase-controller"
        ],
        "operationId": "createRecipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purchases": {
      "get": {
        "tags": [
          "purchase-controller"
        ],
        "operationId": "purchases",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "purchase-controller"
        ],
        "operationId": "receive",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "products",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 40
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageProductView"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/import": {
      "post": {
        "tags": [
          "import-export-controller"
        ],
        "operationId": "upload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/productions": {
      "get": {
        "tags": [
          "purchase-controller"
        ],
        "operationId": "productions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "purchase-controller"
        ],
        "operationId": "produce",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/purchases/{id}": {
      "get": {
        "tags": [
          "purchase-controller"
        ],
        "operationId": "purchase",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/products/export": {
      "get": {
        "tags": [
          "import-export-controller"
        ],
        "operationId": "export_1",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "xlsx"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SupplierInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0
          },
          "document": {
            "type": "string",
            "maxLength": 30,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "maxLength": 40,
            "minLength": 0
          }
        },
        "required": [
          "name"
        ]
      },
      "RecipeInput": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "yield": {
            "type": "number",
            "minimum": 0
          },
          "note": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeItemInput"
            },
            "maxItems": 50,
            "minItems": 0
          }
        },
        "required": [
          "items",
          "productId",
          "yield"
        ]
      },
      "RecipeItemInput": {
        "type": "object",
        "properties": {
          "inputId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "inputId",
          "quantity"
        ]
      },
      "ProductInput": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "barcode": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "unit": {
            "type": "string",
            "pattern": "UN|KG"
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "cost": {
            "type": "number",
            "minimum": 0
          },
          "minimumStock": {
            "type": "number",
            "minimum": 0
          },
          "expiresOn": {
            "type": "string",
            "format": "date"
          },
          "perishable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "cost",
          "minimumStock",
          "name",
          "price",
          "unit"
        ]
      },
      "ProductView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "categoryName": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "cost": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "minimumStock": {
            "type": "number"
          },
          "expiresOn": {
            "type": "string",
            "format": "date"
          },
          "perishable": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "PurchaseInput": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "supplierId": {
            "type": "integer",
            "format": "int64"
          },
          "document": {
            "type": "string",
            "maxLength": 60,
            "minLength": 0
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "payment": {
            "type": "string",
            "pattern": "TERM|CASH"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseItemInput"
            },
            "maxItems": 300,
            "minItems": 0
          },
          "accountId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "document",
          "dueDate",
          "items",
          "payment",
          "requestId",
          "supplierId"
        ]
      },
      "PurchaseItemInput": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "number",
            "minimum": 0
          },
          "unitCost": {
            "type": "number",
            "minimum": 0
          },
          "expiresOn": {
            "type": "string",
            "format": "date"
          },
          "newPrice": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "productId",
          "quantity",
          "unitCost"
        ]
      },
      "ProductionInput": {
        "type": "object",
        "properties": {
          "recipeId": {
            "type": "integer",
            "format": "int64"
          },
          "multiplier": {
            "type": "number",
            "minimum": 0
          },
          "actualYield": {
            "type": "number",
            "minimum": 0
          },
          "expiresOn": {
            "type": "string",
            "format": "date"
          },
          "note": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "expiresOn",
          "multiplier",
          "recipeId"
        ]
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "sorted": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
      "PageProductView": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "last": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductView"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}